refactor(LottoMoney): 수익률 계산 테스트 작성

+++ src/main/java/lotto/domain/vo/LottoMoney.java
@@ -12,7 +12,7 @@ public class LottoMoney {
     }
 
     public double calculateRateOfReturn(int profit) {
-        double rateOfReturn = (double) profit / purchaseAmount;
+        double rateOfReturn = profit / purchaseAmount;
 
         return rateOfReturn;
     }
